Move Footer.svelte into a Footer/ directory to match component convention
Other components (GlobalSideNav, Icon, CommentTree, etc.) already live
in their own directory alongside their story file. Footer.svelte was
still a flat file at the lib root, so this brings it in line.
