+++ frontend/src/lib/Footer/Footer.svelte
